+++ app/src/main/java/com/example/androidcourseshpp/domain/repository/ContactsLocalRepository.kt
@@ -10,15 +10,15 @@ interface ContactsLocalRepository {
     suspend fun addContact(contact: ContactInfo): Result<Unit>
     suspend fun addContacts(contacts: List<ContactInfo>): Result<Unit>
     fun getContacts(): Flow<Result<List<SyncContactInfo>>>
-    suspend fun getContactById(id: Int): Result<SyncContactInfo?>
-    suspend fun deleteContactById(id: Int): Result<Unit>
+    suspend fun getContactById(id: Long): Result<SyncContactInfo?>
+    suspend fun deleteContactById(id: Long): Result<Unit>
     suspend fun clearContacts(): Result<Unit>
-    suspend fun deleteContactsByIds(ids: List<Int>): Result<Unit>
+    suspend fun deleteContactsByIds(ids: List<Long>): Result<Unit>
     fun isDatabaseSynced(): Boolean
     fun setDatabaseSynced(isSynced: Boolean)
-    suspend fun setSyncStateToContact(contactId: Int, syncAction: SyncAction): Result<Unit>
+    suspend fun setSyncStateToContact(contactId: Long, syncAction: SyncAction): Result<Unit>
     suspend fun refreshContacts(
         newContacts: List<ContactInfo>,
-        deletedContactIds: List<Int>
+        deletedContactIds: List<Long>
     ): Result<Unit>
 }
